refactored stairs
